#2634 - fix config file locator
